Refactor AppList page
